Move test classes files to match its target package.
