+++ android-ui/src/androidTest/java/app/boletinhos/widget/currency/QuindimRobot.kt
@@ -1,23 +1,19 @@
-package app.boletinhos.widget
+package app.boletinhos.widget.currency
 
 import android.view.Gravity
 import android.view.KeyEvent
-import android.widget.FrameLayout.LayoutParams
-import android.widget.FrameLayout.LayoutParams.MATCH_PARENT
-import android.widget.FrameLayout.LayoutParams.WRAP_CONTENT
+import android.widget.FrameLayout
 import androidx.test.core.app.ActivityScenario
-import androidx.test.espresso.Espresso.onView
-import androidx.test.espresso.action.ViewActions.pressKey
-import androidx.test.espresso.action.ViewActions.typeText
-import androidx.test.espresso.assertion.ViewAssertions.matches
-import androidx.test.espresso.matcher.ViewMatchers.withId
+import androidx.test.espresso.Espresso
+import androidx.test.espresso.action.ViewActions
+import androidx.test.espresso.assertion.ViewAssertions
+import androidx.test.espresso.matcher.ViewMatchers
 import app.boletinhos.R
 import app.boletinhos.testutil.TestActivity
 import app.boletinhos.testutil.currencyInputHasRawValue
 import app.boletinhos.testutil.inputLayoutHasPrefix
 import app.boletinhos.testutil.textInputHasTextValue
-import app.boletinhos.widget.currency.CurrencyInput
-import org.hamcrest.Matchers.allOf
+import org.hamcrest.Matchers
 import java.util.Locale
 
 /* Android Q @_@ */
@@ -46,34 +42,34 @@ class QuindimRobot(private val scenario: ActivityScenario<TestActivity>) {
     private fun TestActivity.createCurrencyInputView(): CurrencyInput {
         return CurrencyInput(context = rootView.context, locale = locale).apply {
             id = VIEW_ID
-            layoutParams = LayoutParams(MATCH_PARENT, WRAP_CONTENT).apply {
+            layoutParams = FrameLayout.LayoutParams(MATCH_PARENT, WRAP_CONTENT).apply {
                 gravity = Gravity.CENTER
             }
         }
     }
 
-    private fun view() = onView(allOf(withId(VIEW_ID)))
+    private fun view() = Espresso.onView(Matchers.allOf(ViewMatchers.withId(VIEW_ID)))
 
-    private fun inputView() = onView(allOf(withId(INPUT_VIEW_ID)))
+    private fun inputView() = Espresso.onView(Matchers.allOf(ViewMatchers.withId(INPUT_VIEW_ID)))
 
     fun checkIfCurrencySymbolIsShown(expectedSymbol: String) = apply {
-        view().check(matches(inputLayoutHasPrefix(expectedSymbol)))
+        view().check(ViewAssertions.matches(inputLayoutHasPrefix(expectedSymbol)))
     }
 
     fun hasText(text: String) = apply {
-        view().check(matches(textInputHasTextValue(text)))
+        view().check(ViewAssertions.matches(textInputHasTextValue(text)))
     }
 
     fun hasRawValue(value: Long) = apply {
-        view().check(matches(currencyInputHasRawValue(value)))
+        view().check(ViewAssertions.matches(currencyInputHasRawValue(value)))
     }
 
     fun type(text: String) = apply {
-        inputView().perform(typeText(text))
+        inputView().perform(ViewActions.typeText(text))
     }
 
     fun backspace() = apply {
-        inputView().perform(pressKey(KeyEvent.KEYCODE_DEL))
+        inputView().perform(ViewActions.pressKey(KeyEvent.KEYCODE_DEL))
     }
 
     private companion object {
